+++ ui/src/app/modules/plugins/plugins.component.ts
@@ -2,7 +2,7 @@ import { NgClass } from '@angular/common'
 import { Component, ElementRef, inject, OnDestroy, OnInit, ViewChild } from '@angular/core'
 import { FormControl, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { NavigationEnd, Router } from '@angular/router'
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap'
+import {NgbModal, NgbTooltip} from '@ng-bootstrap/ng-bootstrap'
 import { TranslatePipe, TranslateService } from '@ngx-translate/core'
 import { ToastrService } from 'ngx-toastr'
 import { firstValueFrom, Subscription } from 'rxjs'
@@ -27,6 +27,7 @@ import { PluginSupportComponent } from '@/app/modules/plugins/plugin-support/plu
     PluginCardComponent,
     TranslatePipe,
     NgClass,
+    NgbTooltip,
   ],
 })
 export class PluginsComponent implements OnInit, OnDestroy {
@@ -44,6 +45,7 @@ export class PluginsComponent implements OnInit, OnDestroy {
 
   public mainError = false
   public loading = true
+  public tab: 'main' | 'stats' = 'main'
   public installedPlugins: any = []
   public childBridges = []
   public showSearchBar = false
@@ -291,10 +293,23 @@ export class PluginsComponent implements OnInit, OnDestroy {
         this.form.setValue({ query: '' })
         this.loadInstalledPlugins()
       }
-      setTimeout(() => this.searchInput.nativeElement.blur(), 0)
     } else {
+      window.document.querySelector('body').classList.remove('bg-black')
+      this.tab = 'main'
       this.showSearchBar = true
       setTimeout(() => this.searchInput.nativeElement.focus(), 0)
+    }
+  }
+
+  showStats() {
+    if (this.tab === 'stats') {
+      window.document.querySelector('body').classList.remove('bg-black')
+      this.tab = 'main'
+    } else {
+      // Set body bg color
+      window.document.querySelector('body').classList.add('bg-black')
+      this.tab = 'stats'
+      this.showSearchBar = false
     }
   }
 
@@ -306,6 +321,8 @@ export class PluginsComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
+    window.document.querySelector('body').classList.remove('bg-black')
+
     if (this.navigationSubscription) {
       this.navigationSubscription.unsubscribe()
     }